requirements.txt
# Dependências principais
deap
numpy
matplotlib
shapely>=1.8.0

# Otimização e algoritmos genéticos
deap>=1.3.0

# Processamento de dados
pandas>=1.3.0
scikit-learn>=1.0.0

# Processamento paralelo
joblib>=1.1.0

# Arquivos de configuração
PyYAML>=6.0

# Visualização
seaborn>=0.11.0

# Desenvolvimento e testes
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0

# Documentação
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
